Refactor and complete the program*
The program was heavily refactored and now there is code for reading
the resulting matrix from the socket even if it's broken up into
multiple packets. Most likely this program still doesn't work,
as I have no server to test it yet and I won't write mocks and unit
tests for it.