Final bug fixes and minor improvements
Besides fixing bugs, I have added a log of test results in
testing/testing_run.log, as well as the CSV files with data about
when and which messages the clients and the server sent or received.
This can be used to calculate the different response times you
might want to know, and to correlate them with the number of
simultaneous connections during testing or the volume of data being
transmitted or processed.