Rewrite the server using async programming
The server now uses the Tokio runtime and mostly async functions
for most of it's work. The matrix transposition is done on a fixed
thread pool by rayon and integrated into the rest of async code using
tokio::sync::oneshot. This way the matrix transposition is done
on a separate and fixed-size thread pool, which is good for CPU-bound
tasks, and the rest is done asynchronously with Tokio.

Also there are some changes to the application protocol, which will
be updated in README.md soon.